benchmarks/DeathstarBench/get-user-id-sdc/requirements.txt
python-memcached
pymongo
